core: don't add a dependency to dbus.socket for the user bus
When running off dbus-daemon, the user bus is not in a well-defined
location, and is sometimes used as the session bus or something
entirely different. In any case, it is safe to assume the bus already
exists.
When running off kdbus, there is no dbus.socket at all.

This fixes running dbus-activated services in a kdbus environment
from the user manager.